Updated README & default precision value

+++ deepseek_wrapper/deepseek_wrapper.py
@@ -55,7 +55,7 @@ class DeepseekWrapper:
 
     def load_pretrained_model(
         self,
-        precision: Literal["fp32", "fp16", "int8"] = "fp32"
+        precision: Literal["fp32", "fp16", "int8"] = "fp16"
     ) -> bool:
         # Check precision
         if precision not in self._precision_vals:
@@ -99,7 +99,7 @@ class DeepseekWrapper:
         return True
 
     def load_stored_model(self, folder: str,
-                          precision: Literal["fp32", "fp16", "int8"] = "fp32"
+                          precision: Literal["fp32", "fp16", "int8"] = "fp16"
                           ) -> bool:
         # Check precision
         if precision not in self._precision_vals:
