+++ qwen_wrapper/qwen_wrapper.py
@@ -51,7 +51,7 @@ class QwenWrapper:
 
     def load_pretrained_model(
         self,
-        precision: Literal["fp32", "fp16", "int8"] = "fp32"
+        precision: Literal["fp32", "fp16", "int8"] = "fp16"
     ) -> bool:
         # Check precision
         if precision not in self._precision_vals:
@@ -95,7 +95,7 @@ class QwenWrapper:
         return True
 
     def load_stored_model(self, folder: str,
-                          precision: Literal["fp32", "fp16", "int8"] = "fp32"
+                          precision: Literal["fp32", "fp16", "int8"] = "fp16"
                           ) -> bool:
         # Check precision
         if precision not in self._precision_vals:
@@ -151,7 +151,7 @@ class QwenWrapper:
 
     def load_stored_peft_model(
         self, folder: str,
-        precision: Literal["fp32", "fp16", "int8"] = "fp32"
+        precision: Literal["fp32", "fp16", "int8"] = "fp16"
     ) -> bool:
         # Check precision
         if precision not in self._precision_vals:
